+++ tests/v4/test_mappings.py
@@ -1,6 +1,8 @@
+import datetime
 from pathlib import Path
 
 from aio_taginfo.api.v4.key import PrevalentValue
+from aio_taginfo.api.v4.key.chronology import KeyChronology
 from aio_taginfo.api.v4.key.overview import KeyOverview, Response
 from aio_taginfo.api.v4.key.similar import SimilarKey
 from aio_taginfo.api.v4.site.config.geodistribution import SiteConfigGeodistribution
@@ -82,3 +84,12 @@ def test_tags_popular():
     type_adapter = TypeAdapter(Response[list[PopularTag]])
     response = type_adapter.validate_json(response_str, strict=True)
     assert response.data[0].key == "building"
+
+
+def test_key_chronology():
+    test_dir = Path(__file__).resolve().parent
+    data_file = test_dir / "responses" / "key_chronology_highway.json"
+    response_str = data_file.read_text()
+    type_adapter = TypeAdapter(Response[list[KeyChronology]])
+    response = type_adapter.validate_json(response_str, strict=True)
+    assert response.data[0].date == datetime.date(2007, 10, 7)