+++ RELEASES.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 * **Breaking**: Rename `TagInfo*Error` classes to `Taginfo*Error`
+* Implement `/api/v4/key/chronology` endpoint
 
 ## [0.3.0] – 2024-04-27
 * Implement `/api/4/tags/popular` endpoint